created sweep.jl, improved resampling structure
